refactor: refact hooks, stores, reducers path
