+++ client/src/hooks/index.ts
@@ -1,4 +1,3 @@
-export { useFetch } from './base/useFetch';
-export { useIntersect } from './base/useIntersect';
-export { accountReducer, defaultAccountState } from './meta/accountReducer';
-export { eventsReducer, defaultEventsState } from './meta/eventsReducer';
+export { default as useApiRequest } from './useApiRequest';
+export { default as useFetch } from './useFetch';
+export { default as useIntersect } from './useIntersect';
