+++ client/src/stores/accountStore/index.tsx
@@ -10,7 +10,7 @@ import React, {
 import { AccountAction } from 'types/Actions';
 import { AccountState } from 'types/States';
 import { AccountReducer } from 'types/CustomHooks';
-import { accountReducer, defaultAccountState } from 'hooks';
+import accountReducer, { defaultAccountState } from './reducer';
 import { verifyToken, getUserInfo } from 'apis';
 
 export const UserAccountState = createContext<AccountState>(
